display: preserve guest density with fractional scaling
GtkWidget allocations use logical pixels, while QEMU expects framebuffer
dimensions in device pixels. Scale resize requests using the fractional
GdkSurface scale so the guest produces enough pixels for the surface.

Report the corresponding physical dimensions through SetUIInfo as well.
This allows QEMU to expose an EDID density consistent with the host
monitor instead of making the guest display appear artificially larger
and denser.

Round device dimensions consistently with GTK and use the same rounded
extent when snapping Cairo framebuffers. This avoids stretching an extra
row or column across the texture with nearest-neighbor filtering.

Closes: #22

+++ lib/mks-cairo-framebuffer-private.h
@@ -49,6 +49,6 @@ void                 mks_cairo_framebuffer_snapshot   (MksCairoFramebuffer *self
                                                        double               height,
                                                        double               surface_x,
                                                        double               surface_y,
-                                                       int                  scale);
+                                                       double               scale);
 
 G_END_DECLS
